User_Data_Management/server/app/requirements.txt
alembic==1.11.1
apispec==6.3.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.2.0
Flask-Cors==3.0.10       
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4     
flask-smorest==0.42.0    
Flask-SQLAlchemy==3.0.2  
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2      
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
passlib==1.7.4
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
webargs==8.2.0
Werkzeug==2.3.4
